+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Release v0.0.7 (2023-03-07)
+---------------------------
+
+- feat:  add `review` command, now you can review your code changes by using `ai review` command
+- feat:  handle rate limit errors and retry
 
 Release v0.0.6 (2023-03-06)
 ---------------------------